+++ mqtt/src/src/MQTTClient.h
@@ -249,7 +249,7 @@ typedef struct
 	/** The version number of this structure.  Must be 0 */
 	int struct_version;
 	/** The length of the MQTT message payload in bytes. */
-	int payloadlen;
+	unsigned long payloadlen;
 	/** A pointer to the payload of the MQTT message. */
 	void* payload;
 	/**
@@ -858,7 +858,7 @@ DLLExport int MQTTClient_unsubscribeMany(MQTTClient handle, int count, char* con
   * @return ::MQTTCLIENT_SUCCESS if the message is accepted for publication.
   * An error code is returned if there was a problem accepting the message.
   */
-DLLExport int MQTTClient_publish(MQTTClient handle, const char* topicName, int payloadlen, void* payload, int qos, int retained,
+DLLExport int MQTTClient_publish(MQTTClient handle, const char* topicName, unsigned long payloadlen, void* payload, int qos, int retained,
 																 MQTTClient_deliveryToken* dt);
 /**
   * This function attempts to publish a message to a given topic (see also
